+++ packages/networks/lib.esm/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "networks/5.6.1";
+export declare const version = "networks/5.6.2";
 //# sourceMappingURL=_version.d.ts.map