+++ packages/providers/lib.esm/ankr-provider.js
@@ -8,6 +8,12 @@ function getHost(name) {
     switch (name) {
         case "homestead":
             return "rpc.ankr.com/eth/";
+        case "ropsten":
+            return "rpc.ankr.com/eth_ropsten/";
+        case "rinkeby":
+            return "rpc.ankr.com/eth_rinkeby/";
+        case "goerli":
+            return "rpc.ankr.com/eth_goerli/";
         case "matic":
             return "rpc.ankr.com/polygon/";
         case "arbitrum":