+++ packages/providers/src.ts/ankr-provider.ts
@@ -17,8 +17,16 @@ function getHost(name: string): string {
     switch (name) {
         case "homestead":
             return "rpc.ankr.com/eth/";
+        case "ropsten":
+            return "rpc.ankr.com/eth_ropsten/";
+        case "rinkeby":
+            return "rpc.ankr.com/eth_rinkeby/";
+        case "goerli":
+            return "rpc.ankr.com/eth_goerli/";
+
         case "matic":
             return "rpc.ankr.com/polygon/";
+
         case "arbitrum":
             return "rpc.ankr.com/arbitrum/";
     }