+++ packages/providers/src.ts/json-rpc-provider.ts
@@ -57,7 +57,7 @@ function checkError(method: string, error: any, params: any): any {
         const result = spelunk(error);
         if (result) { return result.data; }
 
-        logger.throwError("missing revert data in call exception", Logger.errors.CALL_EXCEPTION, {
+        logger.throwError("missing revert data in call exception; Transaction reverted without a reason string", Logger.errors.CALL_EXCEPTION, {
             error, data: "0x"
         });
     }
@@ -77,28 +77,28 @@ function checkError(method: string, error: any, params: any): any {
     const transaction = params.transaction || params.signedTransaction;
 
     // "insufficient funds for gas * price + value + cost(data)"
-    if (message.match(/insufficient funds|base fee exceeds gas limit/)) {
+    if (message.match(/insufficient funds|base fee exceeds gas limit/i)) {
         logger.throwError("insufficient funds for intrinsic transaction cost", Logger.errors.INSUFFICIENT_FUNDS, {
             error, method, transaction
         });
     }
 
     // "nonce too low"
-    if (message.match(/nonce too low/)) {
+    if (message.match(/nonce (is )?too low/i)) {
         logger.throwError("nonce has already been used", Logger.errors.NONCE_EXPIRED, {
             error, method, transaction
         });
     }
 
     // "replacement transaction underpriced"
-    if (message.match(/replacement transaction underpriced/)) {
+    if (message.match(/replacement transaction underpriced|transaction gas price.*too low/i)) {
         logger.throwError("replacement fee too low", Logger.errors.REPLACEMENT_UNDERPRICED, {
             error, method, transaction
         });
     }
 
     // "replacement transaction underpriced"
-    if (message.match(/only replay-protected/)) {
+    if (message.match(/only replay-protected/i)) {
         logger.throwError("legacy pre-eip-155 transactions not supported", Logger.errors.UNSUPPORTED_OPERATION, {
             error, method, transaction
         });
@@ -566,7 +566,7 @@ export class JsonRpcProvider extends BaseProvider {
         if (method === "call" || method === "estimateGas") {
             const tx = params.transaction;
             if (tx && tx.type != null && BigNumber.from(tx.type).isZero()) {
-                // If there are no EIP-1559 properties, it might be non-EIP-a559
+                // If there are no EIP-1559 properties, it might be non-EIP-1559
                 if (tx.maxFeePerGas == null && tx.maxPriorityFeePerGas == null) {
                     const feeData = await this.getFeeData();
                     if (feeData.maxFeePerGas == null && feeData.maxPriorityFeePerGas == null) {
@@ -671,7 +671,7 @@ export class JsonRpcProvider extends BaseProvider {
         // JSON-RPC now requires numeric values to be "quantity" values
         ["chainId", "gasLimit", "gasPrice", "type", "maxFeePerGas", "maxPriorityFeePerGas", "nonce", "value"].forEach(function(key) {
             if ((<any>transaction)[key] == null) { return; }
-            const value = hexValue((<any>transaction)[key]);
+            const value = hexValue(BigNumber.from((<any>transaction)[key]));
             if (key === "gasLimit") { key = "gas"; }
             result[key] = value;
         });